Enable compiling extensions statically into PHP
Separate header files, split definitions required by internal_functions.c.
