+++ source/pdo_sqlsrv/pdo_init.cpp
@@ -17,7 +17,11 @@
 //  IN THE SOFTWARE.
 //---------------------------------------------------------------------------------------------------------------------------------
 
-#include "php_pdo_sqlsrv.h"
+extern "C" {
+  #include "php_pdo_sqlsrv.h"
+}
+
+#include "php_pdo_sqlsrv_int.h"
 
 #ifdef ZTS
 ZEND_TSRMLS_CACHE_DEFINE();
