+++ source/sqlsrv/init.cpp
@@ -16,7 +16,11 @@
 //  IN THE SOFTWARE.
 //---------------------------------------------------------------------------------------------------------------------------------
 
-#include "php_sqlsrv.h"
+extern "C" {
+  #include "php_sqlsrv.h"
+}
+
+#include "php_sqlsrv_int.h"
 
 #ifdef ZTS
 ZEND_TSRMLS_CACHE_DEFINE();
