+++ source/sqlsrv/stmt.cpp
@@ -18,7 +18,12 @@
 //---------------------------------------------------------------------------------------------------------------------------------
 
 // *** header files ***
-#include "php_sqlsrv.h"
+extern "C" {
+  #include "php_sqlsrv.h"
+}
+
+#include "php_sqlsrv_int.h"
+
 #ifdef _WIN32
 #include <sal.h>
 #endif // _WIN32
